sqrt_c corrected in complex.c
git-svn-id: file:///home/svn/svn/wphase_project/trunk@78 cdf9f54e-c1c6-40d2-88f3-cdd711570eaa

+++ src/complex.c
@@ -170,12 +170,12 @@ conj_c (complex u)
 complex 
 sqrt_c (complex u)
 {
-       complex         w ;
-	   float           arg, norm ; 
-	   arg    = atan2(u.imag, u.real)/2. ;
-	   norm   = sqrt(sqrt(u.real*u.real+u.imag*u.imag)) ;
-	   w.real = norm*cos(arg) ;
-	   w.imag = norm*sin(arg) ;
+       complex w ;
+       double  arg, norm ; 
+       arg    = atan2(u.imag, u.real)/2. ;
+       norm   = sqrt(sqrt(u.real*u.real+u.imag*u.imag)) ;
+       w.real = norm*cos(arg) ;
+       w.imag = norm*sin(arg) ;
        return (w) ;
 }
